Make instance_occurence plaintext-based
Instead of just counting occurrences of the link anchor text in
the raw wikitext, which would introduce all kinds of false
positives, count occurrences in top-level text nodes. This matches
the logic for locating anchor candidates (which is also limited to
top-level text nodes) and is easy to replicate on Parsoid HTML.

Also drop some unneeded imports.

Bug: T271604
Change-Id: I19ee704c903115fc9686d8c03d407b4dc3cefeb3