training-pipeline: Remove link model from tmp if it exists
stat1008 is sometimes used for querying to validate the results of a
finished pipeline. If you use MySQL, it will copy the linkmodel into the
/tmp directory. If you try querying after regenerating the datasets for
a wiki, the old linkmodel remains in place for /tmp; it should be
removed so that subsequent queries use the new one loaded from the MySQL
table.

Bug: T268901
Change-Id: I0dc1253b9f380e2ab4c79dc8ec084bf75e85ad4d