Provide ability to download datasets from remote store
Add a --download option which looks to see if there are datasets to
import from https://analytics.wikimedia.org/published/datasets/one-off/research-mwaddlink/

The script checks if a checksum is stored in the link recommendation
database, and if one is present, it will verify if it differs from the
checksum on the remote dataset store.

If the checksums differ, the script will download the datasets and checksum
files, before proceeding to verify the checksums and load the datasets.

The intended usage for this script is as a cron like process in the
Kubernetes deployed application to periodically fetch and import
datasets for wikis as new datasets are added or updated.

The current iteration of the script requires a specific wiki ID, a
follow-up will add support for operating on a set of wiki IDs.

Bug: T266826
Change-Id: I7b2be16662b4f196140b01bea4438d1269beeebd