load-datasets: Allow download/import of all published datasets
To facilitate the usage of load-datasets.py in a cron-like job in
production, load-datasets.py is modified to parse a listing of wikis
that we have datasets for, and iterate over that list to download/import
datasets as needed.

The publish-datasets.sh script is adjusted to write to a wikis.txt file
to maintain the listing of available datasets, this seemed preferable to
parsing HTML to get the list.

Bug: T272419
Bug: T266826
Change-Id: I6868267c31fabcea9379498e7ace9d6976ae344e