Use random query string to bypass cache
When making HTTP requests for 1) new wikis to import datasets for, 2) checksums to
validate, or 3) datasets to download, it is possible to get a cached
response. We want to bypass the cache so that a dataset is imported as
soon as it is available, but also so we are sure the latest dataset
files are downloaded in the event of a new version of a dataset is
detected via a modified checksum.

Bug: T266826
Change-Id: I86ff7c580a6676895fa007c83d6d521b7f6df42f