Support overriding analytics urls
Per comments in

https://gerrit.wikimedia.org/r/c/operations/deployment-charts/+/666982,

we tend to heavily discourage access to internal resources via the edge
caches, for both maintenability as well as operational reasons. Namely,
we want to avoid using the edge caches for populating the datasets as
this can open a pathway for polluting the edge caches with non organic
traffic, makes debugging a tad more difficult as extra components are
added in the path way of the request and in general makes depooling
datacenters more difficult (that last point doesn't apply to this case
however due to the fact that analytics resources are eqiad only).

Amend the url building code to support passing an enviromental variable
that will allow to direct the instances running in production to the
correct internal resources

Change-Id: I0ef88d7360cd7bd1610931d7ade31bee975e8f4f