Lower max ngram length and implement in process cache
The combined effect of lowering the maximum ngram length is that we
roughly cut the processing time in half for an article like
Lipsko (~4.5 seconds to ~2.3 seconds) with a reduction in queries to the
anchors table from ~11,000 to 7,540. The in process cache access also
saves 2,072 queries on anchors and 371 queries on w2vfiltered.

The tradeoff is that longer link recommendations (about 1% of the total)
will not be recommended.

The patch also implements a profiler when FLASK_DEBUG is enabled.

Bug: T277342
Change-Id: I37ddce14d8cc9332935e8cd68f1c5d9d36a04f3a