Output valid domains to query for on error
The service currently only tells you that it fails to load a model for a
particular wiki domain, but not which wiki domains can be queried. From
Seve's suggestion, instead of adding a new endpoint with this data, for
now we just output the list of available domains to be queried.

Bug: T277288
Change-Id: Ib2c7aaed7fffdcb6d498f552dda2b4e7c0377fe3