Modify URL structure for service
The main changes are:

* API version is set to v1 as it's now live in the API gateway
* First parameter is "project" which is currently not really used, but
may be in the future
* Domain is the second parameter instead of wiki ID

Some hacks were added to translate from the input queries to a value
thath can be used by our dataset loader. In a follow up, we can modify
the pipeline generation script and remaining references to wiki_id to
use a project/domain tuple instead.

Bug: T277288
Bug: T277865
Change-Id: Id9697b8dcc4186cacdcf6ee76286c3dbb7aae6d2