Preload the app in gunicorn
Per T278223 this has led to an increased memory consumption due to the
5x increase in the number of workers. Utilize gunicorn's preload_app
setting to preload the application in gunicorn before spawning the
workers. The downside is that app reload aren't really easy anymore, but
since we are deploying this in a k8s infrastructure where we restart
everything in every code/configuration deployment, this is fine.

Bug: T278223
Change-Id: Icfae18a32722404089cd5828b9ec4c1884c8de53