Allow slashes in titles
Flask's 'string' route parameter type doesn't allow slashes
(and apparently urlencoding is not a reliable way of getting
around that as Flask runs as a WSGI server and as such gets
already decoded URLs). The built-in 'path' type does, but still
disallows leading slashes, so use a custom type instead.

Change-Id: I7ffe4069e41ebf2defb1cad5e39a0ec570816866