Fix parsing error when generating the backtesting data
mwparserfromhell's output (wikicode) can yield a None-type. In this
case, replacing symbols in a string yields an error (this was observed
for ukwiki). This fix moves that part into the try-except statement
so the function will return None. This corresponds to skipping the
parsing of the sentence for the generation of the backtesting.

Tested for ukwiki; the script now runs without errors and yields
expected output.

Bug: T290011
Change-Id: I56fa70b3c7e3f5eb0a177abb7bb3dd13007acc3e