Fix link target location bug and add test
The logic for identifying words to be linked uses a \b-delimited
regex to update the wikitext, but then plain text search to
generate the wikitext offset. When the first occurrence of the
word has other characters attached to it, and the same text node
includes another occurrence with no extra characters, this leads
to the structured output of the algorithm identifying that word,
which leads to Parsoid adding a nowiki tag when the edit is
actually performed.

This is not necessarily an improvement: wiki policy requires the
first occurrence of a word should be linked. But the current
logic already violates that (when the first text node with
occurrences of the word contains only prefixed or suffixed
occurrences, that text node will be skipped because the 'found'
flag will be false) so this patch just makes the behavior more
consistent. Also, differentiating between "foos" being the same
word as "foo" (so we'd want to turn it into "[[foo|foos]]" as
opposed to the behavior before this patch, "[[foo]]s") and
"foosball" (which is a different word and should never link to
"foo") is hard and would require introducing new kinds of
language processing to the algorithm. Dealing with that set of
problems is left for the future.

Depends-On: I82ae7588712d95d6076e5bc7e60fbfa2100d0a84
Bug: T283985
Change-Id: I89d5150b59dffd31d674f734cb48de49e0950ed1