Allow excluding sections from link generation
Why:

* As noted in the task, there are some sections that editors do not want
to see link suggestions generated for, e.g. "References". This patch
allows us to exclude sections from link suggestion generation.

What:

* Define a new paramter to the CLI and API endpoints for "sections to
exclude" from link generation. With CLI, you can repeat the argument to
pass multiple section titles, e.g. flask query --sections-to-exclude=Foo
--sections-to-exclude=Bar. The lead section can be excluded with the
special keyword %LEAD%
* The section titles must be an exact match; no wildcard or regex.
That's done to keep things simple.
* Cap the limit of sections-to-exclude somewhat arbitrarily at 25

How:

* When generating link suggestions, instead of iterating over text
nodes, use mwparserfromhell's get_sections to iterate over each section
of the document, including the lead section. If the node is in the
section to exclude, break out of the loop. Exclusion rules do not flow
downwards, so link suggestions can still be generated for a sub-section
of a section that was excluded. That's done to keep the algorithm as
simple as possible, and also because the expected use case is a couple
of entries for sections like "References" and "Sources" which are
unlikely to have sub-sections anyway.
* A follow-up patch will provide a configuration UI in
Special:EditGrowthConfig where communities can define the list of
section titles to exclude, and those will be read and used by refreshLinkRecommendations.php

Bug: T279519
Change-Id: I2c96bcc77e50af91666ccc7031babe54fb2260f9