Use locale-specific lowercasing
Languages can have specific rules about lowercasing that the default
Python string library doesn't know about. This patch introduces the
PyICU library to handle lowercasing of strings specific to a language
code. The language code is now an optional query parameter in the
request, if not provided, the app will fall back to the "domain" which
in most cases is correct. In cases where it is not correct (e.g. for
"simple" it should be "en" and not "simple"), then PyICU's Locale
library will fall back to 'en'.

Bug: T308244
Change-Id: I962037e614fa5cdd1fce443caf94ce84b7c7b421