pipeline: Publish the test image
Building the test image on arm64 macOS is difficult due to multiple
compatibility issues with the machine learning libraries that we include
in this application.

Publishing the test image makes it easier to run the unit and
integration tests on a pre-built test image.

The resulting image will be available as
docker-registry.wikimedia.org/wikimedia/research-mwaddlink:test

Change-Id: I4d4c3a8d56b8be0be1b87674051ccee13dbd1a20