Update README and fix flask mwaddlink query
Without this commit, flask mwaddlink query insists
to take all information from the request object (which is
not allowed in CLI context), because if revision=0, then
bool(revision)=False (and thus the or clause applies).

It also appears that there is no longer any `flask query`,
only `flask mwaddlink query`.

Change-Id: If8855df8133076afa6f3e51ab5c362f7046d7826