app: Do not convert underscore to dash in wiki_id
This was not needed in its current location, as it
prevents us from using the service for the bat_smg and
be_x_old wikis, which use underscores in their ID names on
the MediaWiki side, and have their data stored in the link
recommendation database using an underscore, and not a dash.

We do however need to convert the underscore to a dash in MediaWikiApi,
in order to be able to construct the correct URL for API requests.

Bug: T320961
Change-Id: I62ac7ecfbafb9951ea268f5c146ab4d63bb85fd8