pipeline: Make it possible to run git inside container
mwaddlink runs `git rev-parse --short HEAD` to detemrine
the application version, which is exposed to API users.
However, git started to insist the repository directory owner
is the same as the process owner, and otherwise refuses to execute.

This is useful on multi-user machines, where this can prevent
an account compromise, but there is no comparable attack scenario
within the application container (an attacker having shell access
within the container would be a significant issue on its own).

git allows users to disable this security check by an user-level
.gitconfig file (command-line params or repository-level config
is not an option). The runuser's home is /home/runuser; copy
a .gitconfig file there.

For some reason, copying from local requires both source
and destination to be relative paths, which is not feasible
in this case (it has to appear as /home/runuser/.gitconfig).
Workaround that by creating a preparation stage in the blubber
file.

While at this, add a syntax directive to make local building
easier.

Bug: T338094
Change-Id: Ice4424dc9bfba1eadbe5d898948e0ef8a98ce8c9