publish-datasets: Require exact match in wikis.txt to skip list update
Why:
Certain wiki IDs contain each other ("gomwiki" and "omwiki" are
both valid wiki IDs). For those projects, wikis.txt was not updated,
leading in missing those wikis when the linkrecommendation k8s service
processes the generated data.

What:
Instead of using only $WIKI as the search pattern for grep,
prefix it with ^ (start of line) and $ (end of line) to require
an exact match, rather than a substring search.

Bug: T340944
Change-Id: Ibaffee6d904bca490f4d39f32108e508c78575f2