tests: Update expected data for provide_query_{get,post}
The models for simplewiki were updated, so slightly different results
are returned. Update the resulting test files to make CI pass.

Replace checksums and application_version in expected_data.json
with "not checked", as those values are ignored by the tests,
and seeing old checksums was confusing during debugging this issue
(and might be in the future).

Bug: T345091
Change-Id: I0e8843fdc9270a91d71fd05327b6f65e94e38c12