docker-compose: Set log level to DEBUG
Why:
While working with mwaddlink locally (ie. when docker-compose is
in use), it is useful to see debug-level log entries. Currently,
the local loglevel is set to WARNING (same in production).

It turns out that since logging.DEBUG et al. are numbers, it is
currently not possible to set them via FLASK_LOGLEVEL yet, as
os.environ.get() always returns a string, and logger.setLevel('10')
fatals.

What:
* Cast FLASK_LOGLEVEL to an int before usage
* Set FLASK_LOGLEVEL to `10` (DEBUG) in docker-compose.yml

Change-Id: I5fd22ca774cbb84a3f96b5550cb95d4a95217ff9