Do not limit the number of sections to exclude
The limit's purpose is unfortunately not documented
in the comment, original commit message or attached
Phabricator tasks. After talking to Kosta Harlan
as the original author of this code, there does not
seem to be any reason to keep it.

Considering GrowthExperiments does not consider
the number of excluded sections to be limited,
and increasing this value would also imply making
GrowthExperiments aware of the (new) limit, let's
select the easier solution and just remove the limit
altogether.

Bug: T380329
Change-Id: I234dad35f611b565cdb33b1ff7115c5f9e3617da